Move unpackResult to separate file for better responsibility separation and adapt seekWorkResponse unit tests to unpackResult.js
